move patch Object.defineProperty to the root /testing folder
